[fix][Step2][피드백 반영] LottoNumber로 책임 이동

+++ src/main/java/lotto/domain/ticket/LottoTickets.java
@@ -15,7 +15,7 @@ public class LottoTickets {
     public WinningTickets findWinning(String[] winningNumbers) {
         return new WinningTickets(
             lottoTickets.stream()
-                        .collect(Collectors.groupingBy(lottoTicket -> (int) lottoTicket.findMatchCount(winningNumbers))));
+                        .collect(Collectors.groupingBy(lottoTicket -> lottoTicket.findMatchCount(winningNumbers))));
     }
 
     public int getLottoTicketsSize() {
